+++ libs/repositories/src/SimulationRepository/tenderlyTypes.ts
@@ -1,18 +1,3 @@
-export interface SimulationInput {
-  input: string;
-  from: string;
-  to: string;
-  value?: string;
-  gas?: number;
-  gas_price?: string;
-}
-
-export interface SimulationData {
-  link: string;
-  status: boolean;
-  id: string;
-}
-
 export interface TenderlyBundleSimulationResponse {
   simulation_results: TenderlySimulation[];
 }
